pyproject.toml
[project]
name = "olympflow"
version = "0.1.0"
description = "Multimodal olympiad problems parsing and clustering"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hdbscan>=0.8.42",
    "httpx>=0.28.1",
    "jupyterlab>=4.5.6",
    "matplotlib>=3.10.8",
    "mistralai>=2.3.1",
    "numpy>=2.4.4",
    "openai>=2.31.0",
    "pillow>=12.2.0",
    "polars>=1.39.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.10",
    "scikit-learn>=1.8.0",
    "tqdm>=4.67.3",
    "umap-learn>=0.5.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
pythonpath = ["src"]
